requirements.txt
gradio==4.25.0
gradio_client==0.15.0
groq==0.4.2
llama-index==0.10.27
llama-index-agent-openai==0.2.2
llama-index-core==0.10.27
llama-index-llms-groq==0.1.3
llama-index-llms-openai==0.1.14
llama-index-vector-stores-qdrant==0.1.6
llama-parse==0.4.0
qdrant-client==1.8.2
torch==2.2.2
transformers==4.39.3
python-dotenv